backend/reservation-service/src/main/java/com/hotel/reservation_service/audit/AuditorAwareImpl.java
package com.hotel.reservation_service.audit;

import jakarta.servlet.http.HttpServletRequest;
import org.springframework.data.domain.AuditorAware;
import org.springframework.stereotype.Component;
import org.springframework.web.context.request.RequestContextHolder;
import org.springframework.web.context.request.ServletRequestAttributes;

import java.util.Optional;

@Component
public class AuditorAwareImpl implements AuditorAware<String> {
    @Override
    public Optional<String> getCurrentAuditor() {
        ServletRequestAttributes attributes = (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();

        if (attributes != null) {
            HttpServletRequest request = attributes.getRequest();

            String userEmail = request.getHeader("Logged-User");
            return Optional.ofNullable(userEmail);
        }
        return Optional.empty();
    }
}
